move shuffle serialization tests to kernel tests
